MM-53226 - Desktop notifications, visibility rules, ringing rules (#452)
* refactor & add hooks; notify when in background (desktop & webapp)

* desktop notifications, ringing rules, simplification of ringing

* finish rebase, mostly channelID -> callID changes

* use main branch calls-common; enable ringingEnabled feature flag

* small bug from switching to callID

* couple more small fixes

* respect user and channel muted settings

* more granular notification preferences (channel + system level)

* ringing should default to on if not set

* use ownerID as caller, hostID -> callerID; fix notification bug

* small fix

* i18n

* MM-53322 - Channel toast fixes (#455)

* record explicit dismisses for channel toast and for later use

* e2e tests

* Translated using Weblate (Japanese) (#457)

Currently translated at 100.0% (161 of 161 strings)

Merge remote-tracking branch 'origin/main'

Translated using Weblate (Korean)

Currently translated at 3.1% (5 of 161 strings)

Translated using Weblate (Czech)

Currently translated at 99.3% (160 of 161 strings)

Translated using Weblate (French)

Currently translated at 60.2% (97 of 161 strings)

Merge remote-tracking branch 'origin/main'

Merge remote-tracking branch 'origin/main'

Translated using Weblate (Polish)

Currently translated at 100.0% (161 of 161 strings)

Translated using Weblate (Dutch)

Currently translated at 100.0% (161 of 161 strings)

Translated using Weblate (Spanish)

Currently translated at 30.4% (49 of 161 strings)

Translated using Weblate (German)

Currently translated at 100.0% (161 of 161 strings)

Update translation files

Updated by "Cleanup translation files" hook in Weblate.

Translated using Weblate (Spanish)

Currently translated at 26.7% (43 of 161 strings)

Translated using Weblate (Czech)

Currently translated at 100.0% (161 of 161 strings)

Translated using Weblate (Chinese (Simplified))

Currently translated at 100.0% (161 of 161 strings)

Merge remote-tracking branch 'origin/main'

Translated using Weblate (German)

Currently translated at 100.0% (161 of 161 strings)

Translated using Weblate (Italian)

Currently translated at 26.0% (42 of 161 strings)

Translated using Weblate (Dutch)

Currently translated at 100.0% (161 of 161 strings)

Translated using Weblate (Korean)

Currently translated at 1.8% (3 of 161 strings)

Translated using Weblate (Croatian)

Currently translated at 96.8% (156 of 161 strings)

Translated using Weblate (Polish)

Currently translated at 100.0% (161 of 161 strings)

Added translation using Weblate (Korean)

Translated using Weblate (Turkish)

Currently translated at 100.0% (161 of 161 strings)

Translated using Weblate (Chinese (Simplified))

Currently translated at 63.9% (103 of 161 strings)

Translated using Weblate (Japanese)

Currently translated at 100.0% (161 of 161 strings)

Merge remote-tracking branch 'origin/main'

Translated using Weblate (Japanese)

Currently translated at 100.0% (159 of 159 strings)

Translated using Weblate (Turkish)

Currently translated at 100.0% (159 of 159 strings)

Translated using Weblate (Japanese)

Currently translated at 1.8% (3 of 159 strings)

Translated using Weblate (Croatian)

Currently translated at 95.5% (152 of 159 strings)

Translated using Weblate (Chinese (Simplified))

Currently translated at 15.7% (25 of 159 strings)

Translated using Weblate (Japanese)

Currently translated at 1.2% (2 of 159 strings)

Translated using Weblate (Uzbek)

Currently translated at 4.4% (7 of 159 strings)

Added translation using Weblate (Uzbek)

Translated using Weblate (Italian)

Currently translated at 20.1% (32 of 159 strings)

Added translation using Weblate (Italian)

Translated using Weblate (Czech)

Currently translated at 100.0% (159 of 159 strings)

Translated using Weblate (Croatian)

Currently translated at 94.9% (151 of 159 strings)

Translated using Weblate (Croatian)

Currently translated at 1.2% (2 of 159 strings)

Added translation using Weblate (Croatian)

Translated using Weblate (German)

Currently translated at 100.0% (159 of 159 strings)

Translated using Weblate (Dutch)

Currently translated at 100.0% (159 of 159 strings)

Translated using Weblate (Polish)

Currently translated at 100.0% (159 of 159 strings)






















Translate-URL: https://translate.mattermost.com/projects/calls/webapp/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/cs/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/de/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/es/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/fr/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/hr/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/it/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/ja/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/ko/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/nl/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/pl/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/tr/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/uz/
Translate-URL: https://translate.mattermost.com/projects/calls/webapp/zh_Hans/
Translation: Calls/webapp


* Translated using Weblate (Korean) (#456)

Currently translated at 100.0% (2 of 2 strings)

Translated using Weblate (Spanish)

Currently translated at 100.0% (2 of 2 strings)

Translated using Weblate (Chinese (Simplified))

Currently translated at 100.0% (2 of 2 strings)

Merge remote-tracking branch 'origin/main'

Added translation using Weblate (Korean)

Merge remote-tracking branch 'origin/main'

Merge remote-tracking branch 'origin/main'

Merge remote-tracking branch 'origin/main'

Translated using Weblate (Japanese)

Currently translated at 100.0% (2 of 2 strings)

Translated using Weblate (Japanese)

Currently translated at 50.0% (1 of 2 strings)

Translated using Weblate (Chinese (Simplified))

Currently translated at 100.0% (2 of 2 strings)

Added translation using Weblate (Uzbek)

Translated using Weblate (Italian)

Currently translated at 100.0% (2 of 2 strings)

Added translation using Weblate (Italian)

Translated using Weblate (Croatian)

Currently translated at 100.0% (2 of 2 strings)

Added translation using Weblate (Croatian)











Translate-URL: https://translate.mattermost.com/projects/calls/standalone/es/
Translate-URL: https://translate.mattermost.com/projects/calls/standalone/hr/
Translate-URL: https://translate.mattermost.com/projects/calls/standalone/it/
Translate-URL: https://translate.mattermost.com/projects/calls/standalone/ja/
Translate-URL: https://translate.mattermost.com/projects/calls/standalone/ko/
Translate-URL: https://translate.mattermost.com/projects/calls/standalone/zh_Hans/
Translation: Calls/standalone


* MM-53261 - Expanded view notifications (#461)

* incoming calls on expanded window; simplify global join

* stack the notifications vertically

* dismiss notification from SwitchCallModal; some formatting

---------


---------

+++ webapp/src/websocket_handlers.ts
@@ -17,11 +17,11 @@ import {
     UserScreenOnOffData,
     UserVoiceOnOffData,
 } from '@calls/common/lib/types';
-import {batchActions} from 'redux-batched-actions';
 
 import {incomingCallOnChannel, removeIncomingCallNotification, userDisconnected} from 'src/actions';
 
 import {JOINED_USER_NOTIFICATION_TIMEOUT, REACTION_TIMEOUT_IN_REACTION_STREAM} from 'src/constants';
+import {notificationSounds} from 'src/webapp_globals';
 
 import {Store} from './types/mattermost-webapp';
 import {
@@ -43,7 +43,7 @@ import {
     VOICE_CHANNEL_CALL_HOST,
     VOICE_CHANNEL_CALL_RECORDING_STATE,
     VOICE_CHANNEL_USER_JOINED_TIMEOUT,
-    CALL_HAS_ENDED,
+    DISMISS_CALL,
 } from './action_types';
 import {
     getProfilesByIds,
@@ -55,6 +55,7 @@ import {
 import {
     connectedChannelID,
     idToProfileInConnectedChannel,
+    ringingEnabled,
     shouldPlayJoinUserSound,
     voiceChannelCalls,
 } from './selectors';
@@ -67,21 +68,17 @@ export function handleCallEnd(store: Store, ev: WebSocketMessage<EmptyData>) {
         window.callsClient?.disconnect();
     }
 
-    const callID = voiceChannelCalls(store.getState())[channelID].ID;
-    store.dispatch(batchActions([
-        {
-            type: VOICE_CHANNEL_CALL_END,
-            data: {
-                channelID,
-            },
+    store.dispatch({
+        type: VOICE_CHANNEL_CALL_END,
+        data: {
+            channelID,
         },
-        {
-            type: CALL_HAS_ENDED,
-            data: {
-                callID,
-            },
-        },
-    ]));
+    });
+
+    if (ringingEnabled(store.getState())) {
+        const callID = voiceChannelCalls(store.getState())[channelID].ID || '';
+        store.dispatch(removeIncomingCallNotification(callID));
+    }
 }
 
 export function handleCallStart(store: Store, ev: WebSocketMessage<CallStartData>) {
@@ -118,13 +115,9 @@ export function handleCallStart(store: Store, ev: WebSocketMessage<CallStartData
         if (channel) {
             followThread(store, channel.id, channel.team_id);
         }
-    } else {
+    } else if (ringingEnabled(store.getState())) {
         // the call that started is not the call we're currently in.
-        const currentUserID = getCurrentUserId(store.getState());
-        if (currentUserID !== ev.data.host_id) {
-            // the call was not started by us.
-            store.dispatch(incomingCallOnChannel(channelID, ev.data.id, ev.data.host_id, ev.data.start_at));
-        }
+        store.dispatch(incomingCallOnChannel(channelID, ev.data.id, ev.data.owner_id, ev.data.start_at));
     }
 }
 
@@ -147,9 +140,10 @@ export async function handleUserConnected(store: Store, ev: WebSocketMessage<Use
         }
     }
 
-    if (userID === currentUserID) {
+    if (ringingEnabled(store.getState()) && userID === currentUserID) {
         const callID = voiceChannelCalls(store.getState())[channelID].ID || '';
         store.dispatch(removeIncomingCallNotification(callID));
+        notificationSounds?.stopRing(); // And stop ringing for _any_ incoming call.
     }
 
     store.dispatch({
@@ -341,4 +335,10 @@ export function handleUserDismissedNotification(store: Store, ev: WebSocketMessa
         return;
     }
     store.dispatch(removeIncomingCallNotification(ev.data.callID));
+    store.dispatch({
+        type: DISMISS_CALL,
+        data: {
+            callID: ev.data.callID,
+        },
+    });
 }
